refactor(agentic-ai): move OAuth client-credentials resolver into http-client, autowire via common config
Relocates OAuthClientCredentialsTokenResolver next to OAuthService in
connector-commons/http-client, drops the resolver's custom OAuthTokenCache
re-registration in favor of autowiring the cache already provided by
ConnectorsAutoConfiguration, and defines the resolver bean itself in that
common config instead of duplicating cache/holder wiring in the
agentic-ai-specific configuration.
