+++ connector-commons/http-client/pom.xml
@@ -78,6 +78,10 @@ limitations under the License.</license.inlineheader>
       <groupId>io.camunda.connector</groupId>
       <artifactId>connector-core</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.jspecify</groupId>
+      <artifactId>jspecify</artifactId>
+    </dependency>
 
     <dependency>
       <groupId>org.mockito</groupId>
