+++ connector-commons/http-client/src/main/java/io/camunda/connector/http/client/authentication/OAuthClientCredentialsTokenResolver.java
@@ -1,15 +1,22 @@
 /*
  * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
- * under one or more contributor license agreements. Licensed under a proprietary license.
- * See the License.txt file for more information. You may not use this file
- * except in compliance with the proprietary license.
+ * under one or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information regarding copyright
+ * ownership. Camunda licenses this file to you under the Apache License,
+ * Version 2.0; you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
-package io.camunda.connector.agenticai.aiagent.chatmodel.provider.authentication.oauth;
+package io.camunda.connector.http.client.authentication;
 
 import io.camunda.connector.api.error.ConnectorException;
-import io.camunda.connector.http.client.authentication.OAuthService;
-import io.camunda.connector.http.client.authentication.OAuthTokenCache;
-import io.camunda.connector.http.client.authentication.TokenResponse;
 import io.camunda.connector.http.client.client.HttpClient;
 import io.camunda.connector.http.client.model.auth.OAuthAuthentication;
 import org.jspecify.annotations.Nullable;
