--- connectors/agentic-ai/connector-agentic-ai/src/main/java/io/camunda/connector/agenticai/aiagent/chatmodel/provider/authentication/oauth/package-info.java
@@ -1,10 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
- * under one or more contributor license agreements. Licensed under a proprietary license.
- * See the License.txt file for more information. You may not use this file
- * except in compliance with the proprietary license.
- */
-@NullMarked
-package io.camunda.connector.agenticai.aiagent.chatmodel.provider.authentication.oauth;
-
-import org.jspecify.annotations.NullMarked;
